Crate samd_timer

samd-timer

This library provides a type-safe API for the Timer/Counter and Timer/Counter for Control peripherals that are present on Microchip SAM series microcontrollers.

Configuration

This library uses a builder pattern to turn a timer configuration into a timer instance. Configurations are not consumed upon building and can be used to configure multiple timer instances.

Clocks

This library uses the clock system provided by the atsamd-hal crate in order to make configuring the clock system as easy as possible.

Example

use target_device::Peripherals;
use atsamd_hal::clock::GenericClockController;
 
let mut peri = Peripherals::take().unwrap();
 
// Configure the HAL clock controller.
let mut clock_ctrl = GenericClockController::with_internal_32kosc(
    peri.GCLK,
    &mut peri.MCLK,
    &mut peri.OSC32KCTRL,
    &mut peri.OSCCTRL,
    &mut peri.NVMCTRL
);

// Assign 120MHz clock to TCC0
let fast_clock = clock_ctrl.gclk0();
let timer_clk = clock_ctrl.tcc0_tcc1(&fast_clock).unwrap();
 
let mut config = TimerConfig::control();
 
config.prescaler(Prescaler::Div64)
    .sync(Synchronisation::Prescaler);
 
let timer = config.tcc0(&mut peri.MCLK, &timer_clk, peri.TCC0);

The last line of the above example shows how you can build an instance of a ControlTimer by using one of the instance specific methods. These methods take care of setting up the relavent clocks for the peripheral instance.

Re-exports

pub use config::TimerConfig;

pub use config::TCC;

pub use config::Count8;

pub use config::Count16;

pub use config::Count32;

pub use control::ControlTimer;

pub use timer::Timer;

Modules

config
Configuration types for Timer and ControlTimer.

control
Timer Counter for Control applications.

timer
Timer Counter.

Structs

TC
Type marker for Timer specific configuration.