Module samd_timer::config

Configuration types for Timer and ControlTimer.

See TimerConfig for more information on how to use the types in this module.

Structs

Channels
Bitfield for channel polarity.

Count8
Marker type for the TC count mode.

Count16
Marker type for the TC count mode.

Count32
Marker type for the TC count mode.

DTIEnable
Dead-time insertion generator bitfield.

TC
Type marker for Timer specific configuration.

TCC
Type marker for ControlTimer specific configuration.

TimerConfig
Common initialisation time configuration for Timer and ControlTimer.

WaveformSelect
Bitfield used to select waveform outputs.

Enums

CaptureMode
Capture modes for Timer.

CaptureTrigger
Capture triggers for Timer.

Direction
Counting directions for timers.

Dither
Dither widths for PWM dithering.

OutputMatrixConfig
Output matrix configuration options.

Prescaler
Set of clock prescalers.

Ramp
Ramp operation modes.

Synchronization
Prescaler and counter synchronisation options.

TimerMode
Bit-width modes for Timer.

TimerWaveGen
Wave generation modes for Timer.

Traits

CountMode
Trait for the different counter widths of the Timer peripheral.