Struct samd_timer::config::TimerConfig

pub struct TimerConfig<KIND> { /* fields omitted */ }

Common initialisation time configuration for Timer and ControlTimer.

Both the TC and TCC peripherals require that some configuration values are set before an instance is enabled. This type represents those values using a Builder pattern. To configure a ControlTimer, create a default configuration using TimerConfig::control(). For a Timer, use either TimerConfig::count8() or TimerConfig::count16(), depending on what bit-width you want the timer to use.

Builder methods do not consume the configuration object, so you can reuse the object to configure or reconfigure other timer instances.

Methods

impl TimerConfig<TCC>

pub fn tcc0(
&self,
mclk: &mut PM,
_clock: &Tcc0Tcc1Clock,
tcc: TCC0
) -> ControlTimer<TCC0>

Build and configure a TCC instance.

pub fn tcc1(
&self,
mclk: &mut PM,
_clock: &Tcc0Tcc1Clock,
tcc: TCC1
) -> ControlTimer<TCC1>

Build and configure a TCC instance.

pub fn tcc2(
&self,
mclk: &mut PM,
_clock: &Tcc2Tc3Clock,
tcc: TCC2
) -> ControlTimer<TCC2>

Build and configure a TCC instance.

impl TimerConfig<TC<Count8>>

pub fn tc3(
&self,
mclk: &mut PM,
_clock: &Tcc2Tc3Clock,
tc: TC3
) -> Timer<TC3, Count8>

pub fn tc4(
&self,
mclk: &mut PM,
_clock: &Tc4Tc5Clock,
tc: TC4
) -> Timer<TC4, Count8>

pub fn tc5(
&self,
mclk: &mut PM,
_clock: &Tc4Tc5Clock,
tc: TC5
) -> Timer<TC5, Count8>

impl TimerConfig<TC<Count16>>

pub fn tc3(
&self,
mclk: &mut PM,
_clock: &Tcc2Tc3Clock,
tc: TC3
) -> Timer<TC3, Count16>

pub fn tc4(
&self,
mclk: &mut PM,
_clock: &Tc4Tc5Clock,
tc: TC4
) -> Timer<TC4, Count16>

pub fn tc5(
&self,
mclk: &mut PM,
_clock: &Tc4Tc5Clock,
tc: TC5
) -> Timer<TC5, Count16>

impl TimerConfig<TC<Count32>>

pub fn tc4_5(
&self,
mclk: &mut PM,
_clock: &Tc4Tc5Clock,
tc: TC4_5
) -> Timer<TC4_5, Count32>

impl<C: CountMode> TimerConfig<TC<C>>

pub fn wave_gen(&mut self, value: TimerWaveGen) -> &mut Self

impl<KIND> TimerConfig<KIND>

pub fn dma_oneshot(&mut self, value: bool) -> &mut Self

Enable or disable the DMA oneshot trigger.

pub fn auto_lock(&mut self, value: bool) -> &mut Self

If set, double buffering will be disabled on overflow, underflow, and re-trigger events.

pub fn sync(&mut self, value: Synchronization) -> &mut Self

Chose which clock source is used to reset or reload the counter.

pub fn run_standby(&mut self, value: bool) -> &mut Self

If set, the timer will continue operation in standby mode.

pub fn prescaler(&mut self, value: Prescaler) -> &mut Self

Select which factor is used for the timer prescaler.

pub fn halt_on_debug(&mut self) -> &mut Self

Halt the timer when the device enters debug mode.

pub fn continue_on_debug(&mut self) -> &mut Self

The timer will continue in debug mode.

impl TimerConfig<TCC>

pub fn control_timer() -> TimerConfig<TCC>

Create a configuration for a ControlTimer.

pub fn dither(&mut self, value: Dither) -> &mut Self

Select how many PWM frames are between dithered values.

pub fn no_fault_on_debug(&mut self) -> &mut Self

Do not generate faults when device is in debug mode.

pub fn fault_on_debug(&mut self) -> &mut Self

Generate a non-recoverable fault when device is in debug mode.

pub fn dti_output_cycles(&mut self, value: (u8, u8)) -> &mut Self

Set the number of cycles for the dead-time generator's high and low side, respectively.

pub fn dti_enable(&mut self, value: DTIEnable) -> &mut Self

Enable the dead-time insertion generator for the given outputs.

pub fn output_config(&mut self, value: OutputMatrixConfig) -> &mut Self

Set the matrix routing mode of the waveform generation outputs.

pub fn output_inversion(&mut self, value: WaveformSelect) -> &mut Self

Select which waveform outputs are inverted.

impl<C: CountMode> TimerConfig<TC<C>>

pub fn count8() -> TimerConfig<TC<Count8>>

Create a configuration for Timer with an 8-bit counter.

pub fn count16() -> TimerConfig<TC<Count16>>

Create a configuration for Timer with a 16-bit counter.

pub fn count32() -> TimerConfig<TC<Count32>>

Create a configuration for Timer with a 32-bit counter.