Trait samd_timer::config::CountMode

pub trait CountMode: Default + Sealed {
    type Size;

    const MODE: TimerMode;

    fn get_count(rb: &TC_RB) -> Self::Size;

    fn set_count(rb: &TC_RB, value: Self::Size);

    fn get_cc0(rb: &TC_RB) -> Self::Size;

    fn set_cc0(rb: &TC_RB, value: Self::Size);

    fn get_cc1(rb: &TC_RB) -> Self::Size;

    fn set_cc1(rb: &TC_RB, value: Self::Size);
}

Trait for the different counter widths of the Timer peripheral.

This is a sealed trait that provides an interface for the Timer implementation. It is used by Count8 and Count16.

Associated Types

type Size

Loading content...
Associated Constants

const MODE: TimerMode

Loading content...
Required methods

fn get_count(rb: &TC_RB) -> Self::Size

fn set_count(rb: &TC_RB, value: Self::Size)

fn get_cc0(rb: &TC_RB) -> Self::Size

fn set_cc0(rb: &TC_RB, value: Self::Size)

fn get_cc1(rb: &TC_RB) -> Self::Size

fn set_cc1(rb: &TC_RB, value: Self::Size)

Loading content...