Struct samd_timer::control::ControlTimer

pub struct ControlTimer<T> { /* fields omitted */ }

ControlTimer is used to work with a TCC peripheral instance.

All configuration options that are not configurable when the instance is enabled are found in the TimerConfig<TCC> type. You can use one of the tccx methods on TimerConfig<TCC> to configure a new instance of ControlTimer, or use the reconfigure method on ControlTimer to reconfigure an existing timer instance.

Timer Width

The size of all timer value accesses are always 32-bit, even if the timer instance is 24-bit or 16-bit wide. As per the reference manual, the upper bytes will be read as zero.

Methods

impl<T> ControlTimer<T> where
T: Deref<Target = RegisterBlock>,

pub fn eject(self) -> T

Reset and eject the TCC instance.

Use if you need the timer control structure back before drop.

pub fn enable(&mut self)

Enable the timer.

pub fn disable(&mut self)

Disable the timer.

pub fn reset(&mut self)

Disable the timer and issue a software reset. This clears out all configuration and resets them to the default.

pub fn reconfigure(&mut self, config: &TimerConfig<TCC>)

Reset and reconfigure the timer.

pub fn stop(&mut self)

Stop the timer counting.

pub fn retrigger(&mut self)

Restart the timer count.

pub fn is_stopped(&self) -> bool

Returns whether the timer is stopped or not.

pub fn enabled_interrupts(&self) -> Interrupts

Get the enabled interrupts for the timer instance.

pub fn enable_interrupts(&mut self, value: Interrupts)

Enable or disable the timer interrupts specified by the given bitfield.

pub fn interrupt_status(&self) -> Interrupts

Get the timer instance's interrupt flags.

pub fn clear_interrupts(&mut self, value: Interrupts)

Clear the timer instance's interrupt flags as specified by the bitfield.

pub fn enable_double_buffering(&mut self)

Enable double buffering for period, pattern, and compare values.

Values in the buffers for each of the listed configurations will be copied to the timer when hardware update conditions occur.

pub fn disable_double_buffering(&mut self)

Disable double buffering for period, pattern, and compare values.

Values in the buffers for each of the listed configurations will NOT be copied to the timer when hardware update conditions occur.

pub fn force_buffer_update(&mut self)

Force the values stored in buffer registers to be written into their counterpart registers.

pub fn set_oneshot(&mut self, oneshot: bool)

Set whether or not the timer is a oneshot timer.

pub fn get_count(&self) -> u32

Get the current value of the counter.

Dithering

Depending on the configured dithering behaviour, the appropriate shifts will be performed on the value before being written to the register.

Synchronisation

This operation requires synchronisation to read the value, which will block until complete.

pub fn set_count(&mut self, value: u32)

Set the value of the timer counter.

Dithering

Depending on the configured dithering behaviour, the appropriate shifts will be performed on the value before being written to the register.

pub fn set_wave_gen(&mut self, value: WaveformGen)

Set the wave generation mode of the timer.

pub fn set_channel_polarity(&mut self, polarity: Channels)

Set the channel polarity using the given channel bitfield.

pub fn set_direction(&mut self, value: Direction)

Set the direction of counting for the timer.

pub fn toggle_direction(&mut self)

Reverse the counting direction of the timer.

pub fn get_direction(&self) -> Direction

Get the counting direction of the timer.

pub fn set_cc(&mut self, index: usize, value: u32)

Set the capture/compare register value for a given channel.

Depending on whether the TCC is in 24-bit or 16-bit mode, the value will be truncated to the correct length.

Dithering

Depending on the configured dithering behaviour, the appropriate shifts will be performed on the value before being written to the register. Use set_cc_dither to set the dither value.

pub fn get_cc(&self, index: usize) -> u32

Get the capture/compare register value for a given channel.

Dithering

Depending on the configured dithering behaviour, the appropriate shifts will be performed on the value after reading from the register. Use get_cc_dither to get the dither value.

pub fn set_cc_dither(&mut self, index: usize, value: u8)

Set the capture/compare register dither for a given channel.

Dithering

Depending on the configured dithering behaviour, the appropriate shifts will be performed on the value before being written to the register. Use set_cc to set the CC value.

pub fn get_cc_dither(&self, index: usize) -> u8

Get the capture/compare register value for a given channel.

Dithering

Depending on the configured dithering behaviour, the appropriate shifts will be performed on the value after reading from the register. Use get_cc to get the CC value.

pub fn set_cc_buffer(&mut self, index: usize, value: u32)

Set the capture/compare buffer register value for a given channel.

Depending on whether the TCC is in 24-bit or 16-bit mode, the value will be truncated to the correct length.

Dithering

Depending on the configured dithering behaviour, the appropriate shifts will be performed on the value before being written to the register. Use set_cc_buffer_dither to set the dither value.

pub fn get_cc_buffer(&self, index: usize) -> u32

Get the capture/compare register buffer value for a given channel.

Dithering

Depending on the configured dithering behaviour, the appropriate shifts will be performed on the value after reading from the register. Use get_cc_buffer_dither to get the dither value.

pub fn set_cc_buffer_dither(&mut self, index: usize, value: u8)

Set the capture/compare buffer register dither for a given channel.

Depending on whether the TCC is in 24-bit or 16-bit mode, the value will be truncated to the correct length.

Dithering

Depending on the configured dithering behaviour, the appropriate shifts will be performed on the value before being written to the register. Use set_cc_buffer_dither to set the dither value.

pub fn get_cc_buffer_dither(&self, index: usize) -> u8

Get the capture/compare register buffer dither for a given channel.

Dithering

Depending on the configured dithering behaviour, the appropriate shifts will be performed on the value after reading from the register. Use get_cc_buffer_dither to get the dither value.

pub fn set_period(&mut self, value: u32)

Set the period value of the timer.

Depending on whether the TCC is in 24-bit or 16-bit mode, the value will be truncated to the correct length.

Dithering

Depending on the configured dithering behaviour, the appropriate shifts will be performed on the value before being written to the register. Use set_period_dither to set the dither value.

pub fn get_period(&self) -> u32

Get the period value of the timer.

Dithering

Depending on the configured dithering behaviour, the appropriate shifts will be performed on the value after reading from the register. Use get_period_dither to get the dither value.

pub fn set_period_buffer(&mut self, value: u32)

Set the period buffer value of the timer.

Depending on whether the TCC is in 24-bit or 16-bit mode, the value will be truncated to the correct length.

Dithering

Depending on the configured dithering behaviour, the appropriate shifts will be performed on the value before being written to the register. Use set_period_buffer_dither to set the dither value.

pub fn get_period_buffer(&self) -> u32

Get the period buffer value of the timer.

Dithering

Depending on the configured dithering behaviour, the appropriate shifts will be performed on the value after reading from the register. Use get_period_buffer_dither to get the dither value.

pub fn set_period_buffer_dither(&mut self, value: u8)

Set the period buffer dither value of the timer.

Dithering

Depending on the configured dithering behaviour, the appropriate shifts will be performed on the value before being written to the register. Use set_period_buffer to set the period buffer value.

pub fn get_period_buffer_dither(&self) -> u8

Get the period buffer dither value of the timer.

Dithering

Depending on the configured dithering behaviour, the appropriate shifts will be performed on the value after reading from the register. Use get_period_buffer to get the period value.

pub fn set_pattern_value(&mut self, value: WaveformSelect)

Set the pattern value of each waveform output as specified by the provided bitfield.

pub fn get_pattern_value(&self) -> WaveformSelect

Get the pattern value of each waveform output as a bitfield.

pub fn set_pattern_enable(&mut self, enable: WaveformSelect)

Set the enable pattern for each waveform output using the provided bitfield.

pub fn get_pattern_enable(&self) -> WaveformSelect

Get the enable pattern for each waveform output as a bitfield.

pub fn set_pattern_value_buffer(&mut self, value: WaveformSelect)

Set the pattern value buffer of each waveform output as specified by the provided bitfield.

pub fn get_pattern_value_buffer(&self) -> WaveformSelect

Get the pattern value buffer of each waveform output as a bitfield.

pub fn set_pattern_enable_buffer(&mut self, enable: WaveformSelect)

Set the pattern enable buffer of each waveform output as specified by the provided bitfield.

pub fn get_pattern_enable_buffer(&self) -> WaveformSelect

Get the pattern enable buffer of each waveform output as a bitfield.