Module samd_timer::timer

Timer Counter.

Re-exports

pub use crate::config::Count8;

pub use crate::config::Count16;

Structs

Interrupts
A bitfield that describes the interrupt flags that a timer can trigger.

TC4_5

Timer
Timer is used to work with an instance of a TC peripheral.