Struct samd_timer::timer::Timer

pub struct Timer<T, C: CountMode> where
    T: Deref<Target = RegisterBlock>,  { /* fields omitted */ }

Timer is used to work with an instance of a TC peripheral.

8-bit Mode

When using Timer<T, Count8> mode, you can access a period and period buffer value (on SAMX5X targets).

Methods

impl<T, C: CountMode> Timer<T, C> where
T: Deref<Target = RegisterBlock>,

pub fn get_count(&self) -> C::Size

Get the current value of the counter.

Synchronisation

This operation requires synchronisation to read the value, which will block until complete.

pub fn get_cc0(&self) -> C::Size

Get the compare/capture value of channel 0.

pub fn get_cc1(&self) -> C::Size

Get the compare/capture value of channel 1.

impl<T, C: CountMode> Timer<T, C> where
T: Deref<Target = RegisterBlock>,

pub fn eject(self) -> T

Reset and eject the TC instance.

Use if you need the timer control structure back before drop.

pub fn enable(&mut self)

Enable the timer.

pub fn disable(&mut self)

Disable the timer.

pub fn reset(&mut self)

Disable and reset the timer.

pub fn reconfigure(&mut self, config: &TimerConfig<TC<C>>)

Reset and reconfigure the timer.

pub fn retrigger(&mut self)

Retrigger the timer.

pub fn stop(&mut self)

Stop the timer.

pub fn is_stopped(&self) -> bool

Return true if the timer is stopped.

pub fn get_oneshot(&self) -> bool

Return true if the timer is in oneshot mode.

pub fn set_oneshot(&mut self, oneshot: bool)

Turn onehsot mode on and off.

pub fn enable_interrupts(&mut self, value: Interrupts)

Enable or disable the timer interrupts specified by the given bitfield.

pub fn interrupt_status(&self) -> Interrupts

Get the timer instance's interrupt flags.

pub fn clear_interrupts(&mut self, value: Interrupts)

Clear the timer instance's interrupt flags as specified by the bitfield.

pub fn set_direction(&mut self, value: Direction)

Set the direction of counting for the timer.

pub fn toggle_direction(&mut self)

Reverse the counting direction of the timer.

pub fn get_direction(&self) -> Direction

Get the counting direction of the timer.

pub fn set_count(&mut self, value: C::Size)

Set the value of the timer counter.

pub fn set_cc0(&mut self, value: C::Size)

Set the compare/capture value of channel 0.

pub fn set_cc1(&mut self, value: C::Size)

Set the compare/capture value of channel 1.

impl<T> Timer<T, Count8> where
T: Deref<Target = RegisterBlock>,

pub fn get_period(&self) -> u8

Get period value

pub fn set_period(&mut self, per: u8)

Set period value